refactor(tui): extract SessionPreview into widgets/preview.rs
Move SessionPreview struct and all its impl methods (load, load_streaming,
parse_event_minimal, styled_line_to_ratatui, to_ratatui_color,
format_duration) into a dedicated preview module. Migrate related tests
(format_duration variants) to the new file.
