+++ src/tui/widgets/mod.rs
@@ -5,9 +5,9 @@
 pub mod file_explorer;
 pub mod file_item;
 pub mod logo;
+pub mod preview;
 
-pub use file_explorer::{
-    FileExplorer, FileExplorerWidget, SessionPreview, SortDirection, SortField,
-};
+pub use file_explorer::{FileExplorer, FileExplorerWidget, SortDirection, SortField};
 pub use file_item::{format_size, FileItem};
 pub use logo::Logo;
+pub use preview::SessionPreview;
